requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
lxml==4.9.3
boto3==1.28.0
requests==2.31.0
python-dateutil==2.8.2